fix conversion

+++ arangod/Cluster/ClusterInfo.cpp
@@ -1200,7 +1200,7 @@ void ClusterInfo::loadPlan() {
   }
 
   auto diff = duration<float, std::milli>(clock::now() - start).count();
-  _lpTotal += diff;
+  _lpTotal += static_cast<uint64_t>(diff);
   _lpTimer.count(diff);
 }
 
@@ -1384,7 +1384,7 @@ void ClusterInfo::loadCurrent() {
   }
 
   auto diff = duration<float, std::milli>(clock::now() - start).count();
-  _lcTotal += diff;
+  _lcTotal += static_cast<uint64_t>(diff);
   _lcTimer.count(diff);
 }
